refactor(tools): 拆分 tool_file_ops 为四个专注文件工具 + 前端修复
重构变化：
- 将 346 行的全能 FileOperationsTool（7 操作、11 字段上帝 schema）
  拆分为 4 个专注工具：
  - file_read：仅读取文件内容（2 字段 schema）
  - file_write：写入/创建文件（3 字段 schema）
  - file_manage：删除/重命名文件、创建目录（operation 枚举）
  - file_search：列出目录内容、搜索文件（operation 枚举）
- tools/base.py：新增 check_path_access() 合并安全两步检查
- tools/__init__.py + tool_extractors.py 同步更新

前端改进：
- FilesBubble.vue：新增 delete_file / rename_file / create_directory 渲染分支
- FilesBubble.vue：修复 items 类型（移除 as any）
- displayNames.ts：补全 file_edit/file_manage/file_search 显示名
- registry.ts：新增 file_manage/file_search 映射

Bug 修复：
- _extract_file_edit read 操作补上缺失的 lines 数组（FileEditBubble 行号显示修复）
- file_write 后端返回真实 line_count（原为字节数）
- 保留 file_operations 向后兼容映射
