+++ web/src/components/tools/_shared/displayNames.ts
@@ -16,6 +16,9 @@ const DISPLAY_NAMES: Record<string, string> = {
   file_write: '写入文件',
   file_list: '文件列表',
   file_operations: '文件操作',
+  file_manage: '文件管理',
+  file_search: '文件搜索',
+  file_edit: '文件编辑',
   syntax_check: '语法检查',
   code_quality: '代码质量',
   unit_test: '单元测试',
